// apps/frontend/src/api/types.ts
export type FrameExtras = {
  glassEngravingPrice: number | null;
  glassPrice: number | null;
  measuredWidthIn: number | null;
  measuredHeightIn: number | null;
  recommendedSizeWidthIn: number | null;
  recommendedSizeHeightIn: number | null;
  framePrice: number | null;
  mountPrice: number | null;
};